+++ src/main/java/wtf/meier/tariff/interpreter/model/rate/Rates.kt
@@ -1,9 +1,8 @@
 package wtf.meier.tariff.interpreter.model.rate
 
-import kotlinx.serialization.Contextual
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import wtf.meier.tariff.interpreter.extension.customSerializer.CurrencySerializer
+import wtf.meier.tariff.interpreter.helper.serializer.CurrencySerializer
 import wtf.meier.tariff.interpreter.model.Interval
 import wtf.meier.tariff.interpreter.model.Price
 import java.util.*
